dwarf/dwarfinfo: Cache CUs, direct parse or search from known units
Maintain a cache of compile units parse and a map of their offsets
similar to the one mainained of DIEs by compile units.

Add the ability to parse a random compile unit when the offset of
the compile unit header is known.

Add the ability to search for a compile unit containing (spanning)
a given refaddr, such as that given by a DIE reference class
attribute, starting from the closest previous compile unit.