dwarf/dwarfinfo: Lookup CU by DIE ref or pointer, DIE by refaddr
Maintain a cache of compile units parse and a map of their offsets
similar to the one mainained of DIEs by compile units.

Add the ability to parse a random compile unit when the offset of
the compile unit header is known.

Add the ability to search for a compile unit containing (spanning)
a given refaddr, such as that given by a DIE reference class
attribute, starting from the closest previous compile unit.

Add a function to lookup a DIE from a dwarfinfo given a refaddr,
looking up the CU if needed.

Add a function to lookup a DIE from a NameLUTEntry obtained from
the .debug_pubnames or .debug_pubtypes sections using the cu_ofs
and die_ofs embedded within.

The internal CU parsing iterator is extended to pick up where
the a previous call left off and to maintain the cache.