dwarf/die: Follow reference class DIE attributes on demand
Using the newly added methods in compile unit and dwarf info, follow
DIE reference attributes to their refernced DIE on demand.

Do this as a DIE method and not at attribute resolution to avoid
reference loops and unnecessary work.

The desired attribute to follow is passed by name and the attribute
form is used to determine the lookup method.

For now raise NotImplemented exception for type signature lookup
and supplemental dwarf object files.