dwarf/compileunit: cleanup cache access
Instead of passing a reference DIE to get_cached_DIE, copy the
stream from the get_top_DIE(), which has the side effect
of maintaining the invariant that the top DIE is in the cache if
anything is in the cache.

Switch to bisect_right and check the previous element for the cache
hit, saving a compare against the list length.

Change variable name child to die, cur_offset to offset, and reword
comments around caching.  Clean up the comments.